update to 5.1

+++ Source/OrdersAbilities/Private/Orders/RTSMoveOrder.cpp
@@ -118,7 +118,7 @@ void URTSMoveOrder::CalculateFormation(int32 UnitCount, const FVector2D Directio
     //
 
     // The formation is a square and this is its edge length in units.
-    const int32 EdgeLengthX = FMath::CeilToInt(FMath::Sqrt(UnitCount));
+    const int32 EdgeLengthX = FMath::CeilToInt(FMath::Sqrt(static_cast<float>(UnitCount)));
     const int32 EdgeLengthY = FMath::CeilToInt(UnitCount / static_cast<float>(EdgeLengthX));
 
     // TODO: Remove this magic numbers.
